Security.UI\AdamMil.Security.UI.csproj: Security.UI\KeyForm.cs: Security.UI\KeyForm.Designer.cs: Security.UI\KeyForm.resx: + Added a form to let the user select a single key from a list
Security.UI\KeyServerForm.Designer.cs:
* Moved some controls around

Security.UI\SignaturesForm.Designer.cs:
* Set KeyPreview to true so that the form can be closed with the escape key

Security\GPG.cs:
* Made DefaultStatusMessageHandler not call OnInvalidPassword if the user hit cancel
* Made SignAndEncrypt throw an exception when the user cancels

+++ Security/GPG.cs
@@ -335,7 +335,11 @@ public class ExeGPG : GPG
         {
           cmd.SendPassword(encryptionOptions.Password, false);
         }
-        else DefaultPromptHandler(promptId, state);
+        else if(!state.Canceled)
+        {
+          DefaultPromptHandler(promptId, state);
+          if(state.Canceled) cmd.Kill(); // kill GPG if the user doesn't give the password, so it doesn't keep asking
+        }
       };
 
       cmd.StatusMessageReceived += delegate(StatusMessage msg)
@@ -363,7 +367,8 @@ public class ExeGPG : GPG
 
     if(!cmd.SuccessfulExit) // if the process wasn't successful, throw an exception
     {
-      if(encryptionOptions != null) throw new EncryptionFailedException(state.FailureReasons);
+      if(state.Canceled) throw new OperationCanceledException();
+      else if(encryptionOptions != null) throw new EncryptionFailedException(state.FailureReasons);
       else throw new SigningFailedException(state.FailureReasons);
     }
   }
@@ -3815,7 +3820,7 @@ public class ExeGPG : GPG
         {
           state.DefaultPassword = null;
         }
-        else
+        else if(!state.Canceled) // don't send the OnInvalidPassword message if a user has simply hit Cancel
         {
           OnInvalidPassword(((BadPassphraseMessage)msg).KeyId);
           state.FailureReasons |= FailureReason.BadPassword;
@@ -4136,9 +4141,12 @@ public class ExeGPG : GPG
         }
         else if(!HandleRevokePrompt(cmd, promptId, reason, null, ref lines, ref lineIndex))
         {
-          DefaultPromptHandler(promptId, state);
-          if(state.Canceled) cmd.Kill(); // if the user didn't give the password, kill GPG to prevent it from asking
-        }                                // again
+          if(!state.Canceled)
+          {
+            DefaultPromptHandler(promptId, state);
+            if(state.Canceled) cmd.Kill(); // kill GPG if the user doesn't give the password, so it doesn't keep asking
+          }
+        }
       };
 
       cmd.Start();
